requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pytesseract
pandas
pdf2image
pillow
